requirements.txt
dagster
dagster-webserver
pandas
openpyxl
python-docx
google-genai
chromadb
python-dotenv
groq
